requirements.txt
asgiref==3.5.2
Django==4.0.6
gunicorn==20.1.0
sqlparse==0.4.2

djangorestframework~=3.13.1